==> scripts_r/1.prep_data.Rmd <==
---
title: "Prep data for IPM"
author: "Clayton T. Lamb"
date: '`r format(Sys.time(), "%d %B, %Y")`'
output: github_document
---

```{r render, eval=FALSE,include=FALSE}
# rmarkdown::render(here::here("jags", 'run_IPM.Rmd'),
#                   output_file =  "README.md")
#
 knitr::purl(input=here::here("scripts_r", '1.prep_data.Rmd'),
                   output =  here::here("scripts_r", '1.prep_data.r'))
```

## Load Data

```{r Load packages and data, results='hide', message=FALSE, warning=FALSE}
library(renv)
library(tidybayes)
library(ggmcmc)
library(hrbrthemes)
library(tidylog)
library(tidyverse)

# Load data ---------------------------------------------------------------

hd <- read.csv(here::here("data/clean/blueprint.csv"))%>%
  filter(herd!="Quintette Full")
hn <- hd %>%
  dplyr::select(herd, herd_num)
sgt <- read.csv(here::here("data/clean/herd_sightability.csv")) %>%
  arrange(herd) %>%
  left_join(hn, by = "herd")%>%
  filter(herd!="Quintette Full")
afs <- read.csv(here::here("data/clean/survival.csv")) %>%
  arrange(herd) %>%
  left_join(hn, by = "herd")%>%
  filter(herd!="Quintette Full")
afr <- read.csv(here::here("data/clean/recruitment.csv")) %>%
  arrange(herd) %>%
  left_join(hn, by = "herd")%>%
  filter(herd!="Quintette Full")
recruit.yr <-read.csv(here::here("data/clean/recruitment_counts.csv")) %>%
  arrange(herd) %>%
  left_join(hn, by = "herd")%>%
  filter(herd!="Quintette Full")
counts <- read.csv(here::here("data/clean/counts.csv")) %>%
  arrange(herd) %>%
  left_join(hn, by = "herd")%>%
  filter(herd!="Quintette Full")
trt <- read.csv(here::here("data/clean/treatments.csv")) %>%
  arrange(herd) %>%
  left_join(hn, by = "herd")%>%
  filter(herd!="Quintette Full")

sexratios <- read.csv(here::here("data/clean/sexratios.csv")) %>%
  arrange(herd) %>%
  left_join(hn, by = "herd")%>%
  filter(herd!="Quintette Full")
sexratio_summary <- read.csv(here::here("data/clean/sexratio_summary.csv"))
surv.yr <- read_csv(here::here("data/clean/survival_yrly.csv"))%>%
  arrange(herd) %>%
  left_join(hn, by = "herd")%>%
  filter(herd!="Quintette Full")

covariates <- read.csv(here::here("data/spatial/covariates_simplified.csv")) %>%
  mutate(herd = case_when(
    herd == "Bearhole Redwillow" ~ "Narraway BC",
    TRUE ~ herd
  )) %>%
  arrange(herd) %>%
  left_join(hn, by = "herd")%>%
  filter(!herd%in%c("Quintette Full", "Scott West"))

demog <- read.csv(here::here("data/clean/demog.csv"))%>%
    mutate(herd = case_when(
    herd == "Bearhole Redwillow" ~ "Narraway BC",
    TRUE ~ herd
  )) %>%
    arrange(herd) %>%
  left_join(hn, by = "herd")%>%
  filter(herd!="Quintette Full")


```

## Line up years so that it all affects animals by March year X count. So summer and spring need to be X+1, and then all need -1
## shift all non-winrer covars to following year so that now winter 2023 covariates affect 2023 counts and R, but 2023 summer, spring, and disturbance of interest affect 2024 count.
```{r filter recovery actions, results='hide', message=FALSE, warning=FALSE}

covar.names <- covariates%>%select(-herd,-year,-herd_num,-ECCC)%>%colnames ##remove herd and year

covar.names_scaled <- paste0(covar.names,"_scale")

covariates <- covariates%>%
  pivot_longer(covar.names)%>%
  mutate(year=case_when(name%in%c("snow.depth","cooler.winter","rain.on.snow")~year,
                   TRUE~year+1))%>% ## winter happens before the annual March counts each year so it lines up, rest need to go to next year to match up
  pivot_wider()
```
  
## Restrict data to only years of demographic data not affected by recovery actions, and
## Restrict data to only years of covariate data
```{r filter recovery actions, results='hide', message=FALSE, warning=FALSE}
#### TREATMENT COMBOS
treatment.combos <- trt %>%
  filter(applied == 1) %>%
  group_by(herd, year) %>%
  summarize(
    trt = paste(
      gsub(" ", "", treatment), # Remove spaces
      collapse = "-" # Add - between treatments
    ),
    .groups = "drop"
  ) %>%
  dplyr::select(herd, year, trt)

yrs_keep <- 1970:2020

remove.trts <- treatment.combos %>%
  filter(trt != "transplant") %>% ## this was coded as years following the transplant and it had no effect in Lamb et al. 2024 Eco Apps, so retain
  distinct(trt) %>%
  pull(trt)

afs <- afs %>%
  left_join(treatment.combos %>% mutate(year - 1), by = c("herd", "year")) %>% ## S data coded as start of year so survival to March 2022 is 2021 (so year-1)
  filter(!trt %in% remove.trts) %>%
  ## remove years during transplant, where pop growing due to directly adding animals
  filter(
    !(herd %in% "Charlotte Alplands" & year %in% c(1984:1991) - 1),
    !(herd %in% "Telkwa" & year %in% c(1997:2010) - 1),
    !(herd %in% "South Selkirks" & year %in% c(1987:1989) - 1),
    !(herd %in% "Purcell South" & year %in% c(2012:2013) - 1)
  ) %>%
  filter(year %in% (yrs_keep))

afr <- afr %>%
  left_join(treatment.combos, by = c("herd", "year")) %>%
  filter(!trt %in% remove.trts) %>%
  ## remove years during transplant, where pop growing due to directly adding animals
  filter(
    !(herd %in% "Charlotte Alplands" & year %in% c(1984:1991)),
    !(herd %in% "Telkwa" & year %in% c(1997:2010)),
    !(herd %in% "South Selkirks" & year %in% c(1987:1989)),
    !(herd %in% "Purcell South" & year %in% c(2012:2013))
  ) %>%
  filter(year %in% yrs_keep)

counts <- counts %>%
  left_join(treatment.combos, by = c("herd", "year")) %>%
  filter(!trt %in% remove.trts) %>%
  ## remove years during transplant, where pop growing due to directly adding animals
  filter(
    !(herd %in% "Charlotte Alplands" & year %in% c(1984:1991)),
    !(herd %in% "Telkwa" & year %in% c(1997:2010)),
    !(herd %in% "South Selkirks" & year %in% c(1987:1989)),
    !(herd %in% "Purcell South" & year %in% c(2012:2013))
  ) %>%
  filter(year %in% yrs_keep)

sexratios <- sexratios %>%
  left_join(treatment.combos, by = c("herd", "year")) %>%
  filter(!trt %in% remove.trts) %>%
  ## remove years during transplant, where pop growing due to directly adding animals
  filter(
    !(herd %in% "Charlotte Alplands" & year %in% c(1984:1991)),
    !(herd %in% "Telkwa" & year %in% c(1997:2010)),
    !(herd %in% "South Selkirks" & year %in% c(1987:1989)),
    !(herd %in% "Purcell South" & year %in% c(2012:2013))
  ) %>%
  filter(year %in% yrs_keep)

surv.yr <- surv.yr%>%
  left_join(treatment.combos, by=c("herd","year"))%>%
  filter(!trt%in%remove.trts)%>%
  ##remove years during transplant, where pop growing due to directly adding animals
  filter(
    !(herd %in% "Charlotte Alplands" & year %in% c(1984:1991)-1),
    !(herd %in% "Telkwa" & year %in% c(1997:2010)-1),
    !(herd %in% "South Selkirks" & year %in% c(1987:1989)-1),
    !(herd %in% "Purcell South" & year %in% c(2012:2013)-1)
  ) %>%
  filter(year %in% (yrs_keep))

sum(surv.yr$time)/365
sum(surv.yr$event)
recruit.yr <- recruit.yr%>%
  left_join(treatment.combos, by=c("herd","year"))%>%
  filter(!trt%in%remove.trts)%>%
  ##remove years during transplant, where pop growing due to directly adding animals
  filter(
    !(herd %in% "Charlotte Alplands" & year %in% c(1984:1991)),
    !(herd %in% "Telkwa" & year %in% c(1997:2010)),
    !(herd %in% "South Selkirks" & year %in% c(1987:1989)),
    !(herd %in% "Purcell South" & year %in% c(2012:2013))
  )%>%
  filter(year%in%yrs_keep)

```

## Prep data for IPM
```{r Prep data for IPM, results='hide', message=FALSE, warning=FALSE}
# Herds and herd number
herds <- unique(hd$herd)
nherd <- nrow(hn)
nsight_grp <- length(unique(hd$sight_grp))

#  Years of study
yrs <- seq(from = min(yrs_keep), to = max(yrs_keep), by = 1)
nyr <- length(yrs)
yr_idx <- seq(from = 1, to = nyr, by = 1)
yr_df <- as.data.frame(cbind(yrs, yr_idx))

# Age classes
nage <- 2

# Restrict year range for trouble shooting
rest_yr <- nyr
ntrt <- length(unique(trt$treatment))


##last year of data due to extirpation
last_per_herd <- demog%>%
  filter(round(totN, 0) < 5) %>% #when <5 AF left.
  group_by(herd) %>%
  filter(yrs == min(yrs) & yrs%in% yrs_keep) %>%
  ungroup()%>%
  select(herd_num,yrs)%>%
    full_join(
    tibble(
      herd_num = 1:nherd,
    ),
    by = "herd_num"
  )%>%
  mutate(yrs=replace_na(yrs,max(yrs_keep)))%>%
  left_join(yr_df, by="yrs")%>%
  arrange(herd_num)



## Covariates
##identify covariates to use
# covar.names <-c("dist.hum",
#                 "dist.hum.es",
#                 "dist.all",
#                 "grow.seas.length",
#                 "snow.depth",
#                 "cooler.winter",
#                 "warm.dry.summer",
#                 "start.spring",
#                 "ros",
#                 "ndvi"
#                 )


## scale all covariates
covariates <- covariates %>%
  mutate(across(covar.names,
         .fns=~scale(.x, center=TRUE)%>%as.numeric(),
         .names="{col}_scale")) %>%
  mutate(year = year) %>%
  left_join(yr_df, by = c("year" = "yrs"))



##prep covar vectors
for(i in 1:length(covar.names)){

temp.vect <- covariates %>%
  filter(year %in% yrs_keep)%>%
  select(herd_num, yr_idx, !!covar.names_scaled[i]) %>%
  arrange(herd_num, yr_idx) %>%
  pivot_wider(names_from = yr_idx, values_from = !!covar.names_scaled[i], values_fill = NA) %>%
  select(-herd_num)

assign(covar.names[i], temp.vect, env=.GlobalEnv)
saveRDS(temp.vect, here::here("data/clean/forIPM/", paste0(covar.names[i],".rds")))

rm(temp.name)
rm(temp.vect)
}







## Recruitment
 recruit.yr.priors <- recruit.yr%>%
   ##with duplicates removed so data in priors and IPM are same.
  ## choose which recruitment survey type to use, hierarchy is OSC>MNKA>OTC
  ## choose which survey month to use if >1, 0>5>9
  group_by(herd,year)%>%
  mutate(type=factor(type,levels=c("OSC","MNKA","OTC")),
        season.num=case_when(season=="spring"~9,
                              season=="fall"~5,
                              season=="winter"~0))%>%
  arrange(herd,year,season.num, type)%>%
  slice(1)%>%
   ##add in count type matrix
      mutate(
         type.otc=case_when(type=="OTC"~1, TRUE~0),
                  type.osc=case_when(type=="OSC"~1, TRUE~0),
                  type.mnka=case_when(type=="MNKA"~1, TRUE~0),
         adults=case_when(adults==0~1, TRUE~adults))%>%
     left_join(yr_df, by = c("year" = "yrs"))%>%
   left_join(covariates%>%select(herd,year,!!covar.names_scaled), by=c("herd", "year"))%>%
   ungroup%>%
   left_join(last_per_herd%>%select(herd_num,year.max=yr_idx), by="herd_num")%>%
   filter(yr_idx<=year.max)%>%
  select(herd=herd_num,year=yr_idx,events=calves,trials=adults,!!covar.names_scaled,season.num,type.otc:type.mnka)
 
 
recruit.yr.ipm <-  recruit.yr.priors%>%
  select(herd,year,events,trials,season.num,type.otc:type.mnka)


nr <- nrow(recruit.yr.ipm)


# Survival

surv.yr.priors <-surv.yr%>%
  left_join(yr_df, by = c("year" = "yrs"))%>%
     left_join(covariates%>%select(herd,year,!!covar.names_scaled)%>%mutate(year=year-1), by=c("herd", "year"))%>% ##S is start of year. In covar, 2023 is what happened over past year, so corresponds to actually S 2022.
  left_join(last_per_herd%>%select(herd_num,year.max=yr_idx), by="herd_num")%>%
   filter(yr_idx<=year.max)%>%
  select(herd=herd_num,year=yr_idx,event,time,!!covar.names_scaled)%>%
  filter(time>0)

surv.yr.ipm <- surv.yr.priors%>%
    group_by(herd,year)%>%
  summarise(events=sum(event),
            trials=sum(time))%>%
  ungroup%>%
  select(herd,year,events,trials)

ns <- nrow(surv.yr.ipm)


# Sightability
# Estimates of 1 are not usable here, not sure why they are left in or how this
# even works given qlogis(1) equals Inf
# We need to transform the uncertainty along with the mean, changes made in JAGS
grp_p <- sgt %>%
  left_join(hd, by = c("herd", "herd_num")) %>%
  group_by(sight_grp) %>%
  mutate(
    n = n(),
    mean_grp_p = mean(est, na.rm = TRUE),
    mean_grp_pvar = 0.1^2, 
    mean_grp_ptau = 1 / (mean_grp_pvar)
  ) %>%
  slice(1) %>%
  as.data.frame() %>%
  dplyr::select(sight_grp, mean_grp_p, mean_grp_pvar, mean_grp_ptau) %>%
  arrange(sight_grp)


mean_grp_p <- grp_p$mean_grp_p
mean_grp_ptau <- grp_p$mean_grp_ptau

sgt_grp_ind <- matrix(nrow = nherd, ncol = nyr)
for (sg in 1:nrow(sgt_grp_ind)) {
  sgt_grp_ind[sg, ] <- hd$sight_grp[hd$herd_num == sg]
}



# Abundance -- separate out rows where no sightability data provided

# Survey counts (have sightability)

cdat <- counts %>%
  dplyr::filter(!is.na(Sightability)) %>%
  mutate(
    herd = herd_num,
    age = NA,
    sex = NA,
    mu = as.integer(count),
    tau = case_when(
      count > 0 ~ 1 / ((count * 0.15)^2),
      count %in% 0 ~ 1
    )
  ) %>%
  # tau = 1/(sd*sd)) %>%
  left_join(yr_df, by = c("year" = "yrs")) %>%
  left_join(last_per_herd%>%select(herd_num,year.max=yr_idx), by="herd_num")%>%
  filter(yr_idx<=year.max)%>%
  dplyr::select(herd, year = yr_idx, age, sex, mu, tau) %>%
  dplyr::filter(year <= rest_yr) %>%
  arrange(herd, year)

nc <- nrow(cdat)


# Sightability data
pdat <- counts %>%
  left_join(hd, by = c("herd", "herd_num")) %>%
  dplyr::filter(!is.na(Sightability)) %>%
  mutate(
    herd = herd_num,
    age = NA,
    sex = NA,
    mu = Sightability,
    tau = 1 / (sd * sd)
  ) %>%
  left_join(yr_df, by = c("year" = "yrs")) %>%
  left_join(last_per_herd%>%select(herd_num,year.max=yr_idx), by="herd_num")%>%
  filter(yr_idx<=year.max)%>%
  dplyr::select(herd, year = yr_idx, age, sex, mu, tau) %>%
  dplyr::filter(year <= rest_yr) %>%
  arrange(herd, year)
# pdat$pdat[pdat$pdat == 1] <- 0.95
pdat$tau[pdat$tau == "Inf"] <- mean(pdat$tau[pdat$tau != "Inf"])
np <- nrow(pdat)

# Minimum Estimates
edat <- counts %>%
  dplyr::filter(is.na(Sightability)) %>%
  left_join(hd, by = c("herd", "herd_num")) %>%
  mutate(
    herd = herd_num,
    age = NA,
    sex = NA,
    mu = as.integer(count),
    tau = case_when(
      count > 0 ~ 1 / ((count * 0.2)^2),
      count %in% 0 ~ 10
    ),
    # tau = 1/(sd*sd),
    sight_grp = sight_grp
  ) %>%
  left_join(yr_df, by = c("year" = "yrs")) %>%
  left_join(last_per_herd%>%select(herd_num,year.max=yr_idx), by="herd_num")%>%
  filter(yr_idx<=year.max)%>%
  dplyr::select(herd, year = yr_idx, age, sex, mu, tau, sight_grp) %>%
  dplyr::filter(year <= rest_yr) %>%
  arrange(herd, year)


ne <- nrow(edat)


# Survey timing: most are in March. Those NOT in March:

survey_timing <- recruit.yr.ipm %>%
  dplyr::select(herd, year, season.num) %>%
  rbind(expand.grid(
    herd = (1:nherd)[!(1:nherd) %in% unique(recruit.yr.ipm$herd)],
    year=case_when(n_distinct(afr$year)!=nyr~
                (1:nyr)[!(1:nyr)%in%unique(afr%>%left_join(yr_df, by = c("year" = "yrs"))%>%pull(yr_idx))],
                TRUE~1),
    season.num = 0
  )) %>%
  tidyr::spread(year, season.num)
survey_timing <- survey_timing[, -1]
survey_timing[is.na(survey_timing)] <- 0


# Calves born in May
# May/Jun survey = Spring survey (age = 1 month) --- offset ==9
# Oct survey = Fall survey (age = 5 months) ---- offset ==5
# Mar survey = winter survey (age = 10 months) ---- offset ==0

month_offset <- survey_timing



##count type matrices
count.otc <- recruit.yr.ipm %>%
  dplyr::select(herd, year, type=type.otc) %>%
  rbind(expand.grid(
    herd = (1:nherd)[!(1:nherd) %in% unique(recruit.yr.ipm$herd)],
    year=case_when(n_distinct(afr$year)!=nyr~
                (1:nyr)[!(1:nyr)%in%unique(afr%>%left_join(yr_df, by = c("year" = "yrs"))%>%pull(yr_idx))],
                TRUE~1),
    type = 0
  )) %>%
  tidyr::spread(year, type)
count.otc <- count.otc[, -1]
count.otc[is.na(count.otc)] <- 0 

count.mnka <- recruit.yr.ipm %>%
  dplyr::select(herd, year, type=type.mnka) %>%
  rbind(expand.grid(
    herd = (1:nherd)[!(1:nherd) %in% unique(recruit.yr.ipm$herd)],
    year=case_when(n_distinct(afr$year)!=nyr~
                (1:nyr)[!(1:nyr)%in%unique(afr%>%left_join(yr_df, by = c("year" = "yrs"))%>%pull(yr_idx))],
                TRUE~1),
    type = 0
  )) %>%
  tidyr::spread(year, type)
count.mnka <- count.mnka[, -1]
count.mnka[is.na(count.mnka)] <- 0 

count.osc <- recruit.yr.ipm %>%
  dplyr::select(herd, year, type=type.osc) %>%
  rbind(expand.grid(
    herd = (1:nherd)[!(1:nherd) %in% unique(recruit.yr.ipm$herd)],
    year=case_when(n_distinct(afr$year)!=nyr~
                (1:nyr)[!(1:nyr)%in%unique(afr%>%left_join(yr_df, by = c("year" = "yrs"))%>%pull(yr_idx))],
                TRUE~1),
    type = 1
  )) %>%
  tidyr::spread(year, type)
count.osc <- count.osc[, -1]
count.osc[is.na(count.osc)] <- 0 
count.osc[count.mnka==0 & count.osc==0 & count.otc==0] <-1




# # Starting population size of each herd
# take average of all counts that occured within 3 years of first count per herd
#   Using the "Est_CL" column
# Starting population size of each herd
meancount <- counts %>%
  arrange(herd_num, year) %>%
  group_by(herd_num) %>%
  mutate(keep = year - min(year)) %>%
  dplyr::filter(keep < 4) %>%
  mutate(n1_mean = mean(Est_CL, na.rm = TRUE)) %>%
  slice(1) %>%
  as.data.frame() %>%
  dplyr::select(herd, herd_num, n1_mean) %>%
  right_join(hd, by = c("herd", "herd_num")) %>%
  mutate(n1_mean = ifelse(is.na(n1_mean), mean(n1_mean, na.rm = TRUE), n1_mean)) %>%
  arrange(herd_num)
n1s <- meancount$n1_mean

## starting age class distribution (sensitivity tested this in McNay et al. 2022 and it doesnt have a meaningful effect even if substantially changed)
n1 <- matrix(NA, nrow = nherd, ncol = nage)
for (h in 1:nherd) {
  n1[h, 1] <- (n1s[h] * 0.5) * 0.15
  n1[h, 2] <- (n1s[h] * 0.64) * 0.85
}


# Sex ratio
srdat <- sexratios %>%
  mutate(
    herd = herd_num,
    mu = sratio,
    tau = 1 / (sratio.sd * sratio.sd)
  ) %>%
  left_join(yr_df, by = c("year" = "yrs")) %>%
  left_join(last_per_herd%>%select(herd_num,year.max=yr_idx), by="herd_num")%>%
  filter(yr_idx<=year.max)%>%
  dplyr::filter(!is.na(mu)) %>%
  dplyr::filter(mu != "Inf") %>%
  dplyr::select(herd, year = yr_idx, mu, tau) %>%
  dplyr::filter(year <= rest_yr) %>%
  arrange(herd, year) %>%
  distinct(herd, year, .keep_all = TRUE)

nsr <- nrow(srdat)

first_per_herd <-rbind(cdat%>%select(herd,year),
          edat%>%select(herd,year),
          surv.yr.ipm%>%select(herd,year),
          recruit.yr.ipm%>%select(herd,year))%>%
  rename(herd_num=herd)%>%
  group_by(herd_num) %>%
  summarize(yr = min(year)) %>% ## year before trt starts
  full_join(
    tibble(
      herd_num = 1:nherd,
    ),
    by = "herd_num"
  )
#first_per_herd <- tibble(herd_num=1:nherd,yr=1)



  


#  First year of treatment for each herd
treatment_start <- trt %>%
  filter(applied == 1 & !treatment %in% c("transplant", "reduce moose")) %>%
  group_by(herd_num) %>%
  summarize(yr = min(year) - 1) %>% ## year before trt starts
  full_join(
    tibble(
      herd_num = 1:nherd,
      yr_fill = 2021
    ),
    by = "herd_num"
  ) %>%
  left_join(hd %>% dplyr::select(name = herd, herd_num)) %>%
  mutate(yrs = case_when(
    is.na(yr) ~ yr_fill,
    name %in% "Columbia North" ~ 2004, ## CN reduce moose did work so keep this one
    TRUE ~ yr
  )) %>%
  left_join(yr_df) %>%
  arrange(herd_num) %>%
  pull(yr_idx)


nyr <- rest_yr



## single digit params
single.params <- tibble(
  nherd = nherd,
  nage=2,
  nyr = nyr,
  nc = nc,
  ne = ne,
  ns = ns,
  nr = nr,
  nsr = nsr,
  nsight_grp=nsight_grp,
)
```

## Alternate survival parameritization for JW
```{r S for JW, results='hide', message=FALSE, warning=FALSE}

# Takes a data.frame 'df' with columns c('id','start','end','dead')
## transform into daily survival
source(here::here("other", "seasonalsurvival_fn.r"))
surv.day <- surv.yr%>%
  mutate(id=paste(id, herd, sep="_"))%>%
  filter(time>0)%>%
  dplyr::select(id, herd, DateEntry.bio=enter.date, DateExit.bio=exit.date, event) %>%
  mutate(id_period = paste0(id, DateEntry.bio)) %>%
  ungroup() %>%
  stretch_survival_data("1 day") %>%
  mutate(year = year(DateEntry.bio),
         day=(DateEntry.bio-min(DateEntry.bio))%>%as.numeric(),
         alive=abs(1-dead))%>%
  left_join(hd%>%select(herd,herd_num), by="herd")

surv.day.wide <- surv.day%>%
  arrange(day)%>%
  select(id,herd=herd_num,day,alive)%>%
  pivot_wider(names_from=day, values_from=alive)


```

## Join S and R to covariates for data exploration
```{r join SR, results='hide', message=FALSE, warning=FALSE}
sdat.plot <- read_csv(here::here("data/clean/survival.csv"))%>%
  left_join(covariates%>%
              mutate(year=year-1)%>%
              select(herd,year,!!covar.names, !!covar.names_scaled),
            by=c("herd","year"))%>%
  pivot_longer(c(covar.names, covar.names_scaled))%>%
  mutate(value=as.numeric(value))

rdat.plot <- read_csv(here::here("data/clean/recruitment.csv"))%>%
  left_join(covariates%>%
              select(herd,year,!!covar.names, !!covar.names_scaled),
            by=c("herd","year"))%>%
  pivot_longer(c(covar.names, covar.names_scaled))%>%
  mutate(value=as.numeric(value),
         season=case_when(season_int==1~10,
                          season_int==2~5,
                          season_int==3~1))

```

## Save data for IPM

```{r Save data for IPM, results='hide', message=FALSE, warning=FALSE}
## demographic data
write_csv(cdat, here::here("data/clean/forIPM/cdat.csv"))
write_csv(edat, here::here("data/clean/forIPM/edat.csv"))
write_csv(srdat, here::here("data/clean/forIPM/srdat.csv"))
write_csv(pdat, here::here("data/clean/forIPM/pdat.csv"))
write_csv(surv.yr.priors, here::here("data/clean/forIPM/surv.yr.priors.csv"))
write_csv(surv.yr.ipm, here::here("data/clean/forIPM/surv.yr.ipm.csv"))
write_csv(recruit.yr.priors, here::here("data/clean/forIPM/recruit.yr.priors.csv"))
write_csv(recruit.yr.ipm, here::here("data/clean/forIPM/recruit.yr.ipm.csv"))
write_csv(counts, here::here("data/clean/forIPM/counts.csv"))
write_csv(surv.day.wide, here::here("data/clean/forIPM/surv.day.wide.unique.csv"))

## covariates
write_csv(covariates, here::here("data/clean/forIPM/covariates_scaled.csv"))
saveRDS(covar.names, here::here("data/clean/forIPM/covariate_names.rds"))


## other
write_csv(first_per_herd, here::here("data/clean/forIPM/first_per_herd.csv"))
write_csv(last_per_herd, here::here("data/clean/forIPM/last_per_herd.csv"))
write_csv(month_offset, here::here("data/clean/forIPM/month_offset.csv"))
write_csv(grp_p, here::here("data/clean/forIPM/grp_p.csv"))
write_csv(single.params, here::here("data/clean/forIPM/single.params.csv"))
write_csv(meancount, here::here("data/clean/forIPM/meancount.csv"))
write_csv(yr_df, here::here("data/clean/forIPM/yr_df.csv"))
write_csv(count.otc, here::here("data/clean/forIPM/count.otc.csv"))
write_csv(count.osc, here::here("data/clean/forIPM/count.osc.csv"))
write_csv(count.mnka, here::here("data/clean/forIPM/count.mnka.csv"))
write_csv(hd, here::here("data/clean/forIPM/blueprint.csv"))




##plot data
write_csv(rdat.plot, here::here("data/clean/forIPM/rdat.plot.csv"))
write_csv(sdat.plot, here::here("data/clean/forIPM/sdat.plot.csv"))
 

```
