==> scripts_r/2.run_IPM.Rmd <==
---
title: "BC AB Caribou IPM Climate-Disturbance run"
author: "Clayton T. Lamb"
date: '`r format(Sys.time(), "%d %B, %Y")`'
output: github_document
---


```{r render,include=FALSE, eval=FALSE}
rmarkdown::render(here::here("scripts_r", "2.run_IPM.Rmd"),
  output_file = "README.md"
)
```


## Load Data
```{r Load packages and data, results='hide', message=FALSE, warning=FALSE}
## Load Packages
library(tidybayes)
library(ggmcmc)
library(hrbrthemes)
library(tidylog)
library(tidyverse)

# Load data ---------------------------------------------------------------
hd <- read.csv(here::here("data/clean/forIPM/blueprint.csv"))
hn <- hd %>%
  dplyr::select(herd, herd_num)

## demographic data
cdat <- read_csv(here::here("data/clean/forIPM/cdat.csv"))
nrow(cdat)
edat <- read_csv(here::here("data/clean/forIPM/edat.csv"))
nrow(edat)
sdat <- read_csv(here::here("data/clean/forIPM/surv.yr.ipm.csv"))
nrow(sdat)
rdat <- read_csv(here::here("data/clean/forIPM/recruit.yr.ipm.csv"))
nrow(rdat)
srdat <- read_csv(here::here("data/clean/forIPM/srdat.csv"))
nrow(srdat)
pdat <- read_csv(here::here("data/clean/forIPM/pdat.csv"))
nrow(pdat)
sexratio_summary <- read.csv(here::here("data/clean/sexratio_summary.csv"))

## covariates
# library(ggcorrplot)
# covariates%>%
#   select(!!covar.names)%>%
#   drop_na()%>%
#   cor()%>%
# ggcorrplot(hc.order = TRUE, type = "lower",
#    lab = TRUE)
## start spring and growing season length correlated at -0.72. Just use start spring

dist.hum <- readRDS(here::here("data/clean/forIPM/dist.hum.es.rds"))
snow.depth <- readRDS(here::here("data/clean/forIPM/snow.depth.rds"))
cooler.winter <- readRDS(here::here("data/clean/forIPM/cooler.winter.rds"))
start.spring <- readRDS(here::here("data/clean/forIPM/start.spring.rds"))
warm.dry.summer <- readRDS(here::here("data/clean/forIPM/warm.dry.summer.rds"))
rain.on.snow <- readRDS(here::here("data/clean/forIPM/rain.on.snow.rds"))

## fill 1970's covariate values for rain on snow
early.fill <- rain.on.snow[, 11:16] %>%
  rowwise() %>%
  mutate(m = mean(`11`:`16`))
rain.on.snow[, 1:10] <- early.fill$m

## other
first_per_herd <- read_csv(here::here("data/clean/forIPM/first_per_herd.csv"))
last_per_herd <- read_csv(here::here("data/clean/forIPM/last_per_herd.csv"))
month_offset <- read_csv(here::here("data/clean/forIPM/month_offset.csv"))
grp_p <- read_csv(here::here("data/clean/forIPM/grp_p.csv"))
single.params <- read_csv(here::here("data/clean/forIPM/single.params.csv"))
meancount <- read_csv(here::here("data/clean/forIPM/meancount.csv"))
count.otc <- read_csv(here::here("data/clean/forIPM/count.otc.csv"))
count.osc <- read_csv(here::here("data/clean/forIPM/count.osc.csv"))
count.mnka <- read_csv(here::here("data/clean/forIPM/count.mnka.csv"))
yr_df <- read_csv(here::here("data/clean/forIPM/yr_df.csv"))
```

## Estimate Prior
```{r Estimate Priors, results='hide', message=FALSE, warning=FALSE}
mean.s <- 1-(sum(sdat$events) / sum(sdat$trials))
mean.r <- mean(rdat$events / rdat$trials)

ipm.prior.dat <- list(
  mean.s = mean.s,
  mean.r = mean.r,
  ns = nrow(sdat),
  nr = nrow(rdat),
  sdat = sdat,
  rdat = rdat,
  dist = dist.hum,
  snow.depth = snow.depth,
  cooler.winter=cooler.winter,
  start.spring = start.spring,
  warm.dry.summer = warm.dry.summer,
  rain.on.snow = rain.on.snow
)



prior.mod <-
  "model{
    # priors
	meanS <- logit(mean.s)
	muS ~ dnorm(meanS, 1)
    beta.dist.s ~ dunif(-10,10)
    beta.snow.depth.s ~ dunif(-10,10)
    beta.cooler.winter.s ~ dunif(-10,10)
    beta.start.spring.s ~ dunif(-10,10)
    beta.warm.dry.summer.s ~ dunif(-10,10)
    beta.rain.on.snow.s ~ dunif(-10,10)

	meanR <- logit(mean.r)
	muR ~ dnorm(meanR, 1)
    beta.dist.r ~ dunif(-10,10)
    beta.snow.depth.r ~ dunif(-10,10)
    beta.cooler.winter.r ~ dunif(-10,10)
    beta.start.spring.r ~ dunif(-10,10)
    beta.warm.dry.summer.r ~ dunif(-10,10)
    beta.rain.on.snow.r ~ dunif(-10,10)
    beta.season.r ~ dunif(-10,10)
    beta.type.otc.r ~ dunif(-10,10)
    beta.type.osc.r ~ dunif(-10,10)
    beta.type.mnka.r ~ dunif(-10,10)

    # likelihood
    for (i in 1:ns){
            sdat[i,3] ~ dbin((1-s[i]), sdat[i,4])
            logit(s[i]) <- muS +
            beta.dist.s*dist[sdat[i,1],sdat[i,2]] +
            beta.snow.depth.s*snow.depth[sdat[i,1],sdat[i,2]] +
            beta.cooler.winter.s*cooler.winter[sdat[i,1],sdat[i,2]] +
            beta.start.spring.s*start.spring[sdat[i,1],sdat[i,2]] +
            beta.warm.dry.summer.s*warm.dry.summer[sdat[i,1],sdat[i,2]] +
            beta.rain.on.snow.s*rain.on.snow[sdat[i,1],sdat[i,2]]
    }

        for (i in 1:nr){
            rdat[i,3] ~ dbin(r[i], rdat[i,4])
            logit(r[i]) <- muR +
            beta.dist.r*dist[rdat[i,1],rdat[i,2]] +
            beta.snow.depth.r*snow.depth[rdat[i,1],rdat[i,2]] +
            beta.cooler.winter.r*cooler.winter[rdat[i,1],rdat[i,2]] +
            beta.start.spring.r*start.spring[rdat[i,1],rdat[i,2]] +
            beta.warm.dry.summer.r*warm.dry.summer[rdat[i,1],rdat[i,2]] +
            beta.rain.on.snow.r*rain.on.snow[rdat[i,1],rdat[i,2]] +
            beta.season.r*rdat[i,5]+
            beta.type.otc.r*rdat[i,6] + beta.type.osc.r*rdat[i,7] + beta.type.mnka.r*rdat[i,8]

    }

}"


nth <- 50
nbu <- 0
nch <- 3
nad <- 5000
nit <- 30000

out.priors <- jagsUI::jags(ipm.prior.dat,
  inits = NULL,
  parameters.to.save = c(
    "s", "muS", "beta.dist.s", "beta.snow.depth.s", "beta.cooler.winter.s", "beta.start.spring.s", "beta.warm.dry.summer.s", "beta.rain.on.snow.s",
    "r", "muR", "beta.dist.r", "beta.snow.depth.r", "beta.cooler.winter.r", "beta.start.spring.r", "beta.warm.dry.summer.r", "beta.rain.on.snow.r", "beta.season.r"
  ),
  model.file = textConnection(prior.mod),
  n.chains = nch,
  n.cores = nch,
  n.iter = nit,
  n.burnin = nbu,
  n.thin = nth,
  n.adapt = nad
)

# mcmcplots::mcmcplot(out.priors$samples, par = c("muS", "beta.dist.s", "beta.snow.depth.s", "beta.start.spring.s",  "beta.warm.dry.summer.s", "beta.rain.on.snow.s", "muR", "beta.dist.r", "beta.snow.depth.r", "beta.start.spring.r", "beta.warm.dry.summer.r", "beta.rain.on.snow.r", "beta.season.r"))


priors <- out.priors %>%
  gather_draws(
    muS, beta.dist.s, beta.snow.depth.s, beta.cooler.winter.s, beta.start.spring.s, beta.warm.dry.summer.s, beta.rain.on.snow.s,
    muR, beta.dist.r, beta.snow.depth.r, beta.cooler.winter.r, beta.start.spring.r, beta.warm.dry.summer.r, beta.rain.on.snow.r, beta.season.r
  ) %>%
  group_by(.variable) %>%
  summarise(
    mean = mean(.value),
    sd = sd(.value),
    tau = (1 / (sd^2))
  )
```

## Prep data for IPM
```{r Prep data for IPM, results='hide', message=FALSE, warning=FALSE}
# Herds and herd number
nherd <- single.params$nherd
nsight_grp <- single.params$nsight_grp
nage <- single.params$nage

#  Years of study
nyr <- single.params$nyr

# sightability info
mean_grp_p <- grp_p$mean_grp_p
mean_grp_ptau <- grp_p$mean_grp_ptau

sgt_grp_ind <- matrix(nrow = nherd, ncol = nyr)
for (sg in 1:nrow(sgt_grp_ind)) {
  sgt_grp_ind[sg, ] <- hd$sight_grp[hd$herd_num == sg]
}


## starting age class distribution (sensitivity tested this in McNay et al. 2022 and it doesnt have a meaningful effect even if substantially changed)
n1s <- meancount$n1_mean
n1 <- matrix(NA, nrow = nherd, ncol = nage)
for (h in 1:nherd) {
  n1[h, 1] <- (n1s[h] * 0.5) * 0.15
  n1[h, 2] <- (n1s[h] * 0.64) * 0.85
}

# n1error <- matrix(NA, nrow = nherd, ncol = nage)
# for (h in 1:nherd) {
#   n1error <- 1/((n1)^2)
# }

meansr <- array(NA, c(1, 2))
meansr[1, 1] <- sexratio_summary[1, 1]
# meansr[1,2] <- 1/(sexratio_summary[1,2]^2)
meansr[1, 2] <- 1 / (0.01^2) ## make sex ratio very precise so model doesn't eat as much variance up here

# ## first year of data (legacy...set at 1 so all herds start at year 1 now)
# first <- rep(1, nherd)
first <- first_per_herd%>%pull(yr)
last <- last_per_herd%>%pull(yr_idx)


## Gather data inputs in a list

ipm_dat <- list(
  mean.s = priors %>% filter(.variable == "muS") %>% pull(mean),
  mean.s.tau = priors %>% filter(.variable == "muS") %>% pull(tau),
  mean.r = priors %>% filter(.variable == "muR") %>% pull(mean),
  mean.r.tau = priors %>% filter(.variable == "muR") %>% pull(tau),
  nherd = nherd,
  nyr = nyr,
  first = first,
  last = last,
  month_offset = month_offset,
  count.otc = count.otc,
  count.osc = count.osc,
  count.mnka = count.mnka,
  nc = single.params$nc,
  ne = single.params$ne,
  ns = single.params$ns,
  nr = single.params$nr,
  nsr = single.params$nsr,
  nsight_grp = nsight_grp,
  sight_grp = hd$sight_grp,
  mean_grp_p = mean_grp_p,
  mean_grp_ptau = mean_grp_ptau,
  meansr = meansr,
  n1 = n1,
  cdat = cdat,
  edat = edat,
  sdat = sdat,
  srdat = srdat,
  pdat = pdat,
  rdat = rdat,
  dist = dist.hum,
  snow.depth = snow.depth,
  cooler.winter=cooler.winter,
  start.spring = start.spring,
  warm.dry.summer = warm.dry.summer,
  rain.on.snow = rain.on.snow,
  prior.dist.s.mean = priors %>% filter(.variable == "beta.dist.s") %>% pull(mean),
  prior.dist.s.tau = priors %>% filter(.variable == "beta.dist.s") %>% pull(tau),
  prior.snow.depth.s.mean = priors %>% filter(.variable == "beta.snow.depth.s") %>% pull(mean),
  prior.snow.depth.s.tau = priors %>% filter(.variable == "beta.snow.depth.s") %>% pull(tau),
  prior.cooler.winter.s.mean = priors %>% filter(.variable == "beta.cooler.winter.s") %>% pull(mean),
  prior.cooler.winter.s.tau = priors %>% filter(.variable == "beta.cooler.winter.s") %>% pull(tau),
  prior.start.spring.s.mean = priors %>% filter(.variable == "beta.start.spring.s") %>% pull(mean),
  prior.start.spring.s.tau = priors %>% filter(.variable == "beta.start.spring.s") %>% pull(tau),
  prior.warm.dry.summer.s.mean = priors %>% filter(.variable == "beta.warm.dry.summer.s") %>% pull(mean),
  prior.warm.dry.summer.s.tau = priors %>% filter(.variable == "beta.warm.dry.summer.s") %>% pull(tau),
  prior.rain.on.snow.s.mean = priors %>% filter(.variable == "beta.rain.on.snow.s") %>% pull(mean),
  prior.rain.on.snow.s.tau = priors %>% filter(.variable == "beta.rain.on.snow.s") %>% pull(tau),
  prior.dist.r.mean = priors %>% filter(.variable == "beta.dist.r") %>% pull(mean),
  prior.dist.r.tau = priors %>% filter(.variable == "beta.dist.r") %>% pull(tau),
  prior.snow.depth.r.mean = priors %>% filter(.variable == "beta.snow.depth.r") %>% pull(mean),
  prior.snow.depth.r.tau = priors %>% filter(.variable == "beta.snow.depth.r") %>% pull(tau),
  prior.cooler.winter.r.mean = priors %>% filter(.variable == "beta.cooler.winter.r") %>% pull(mean),
  prior.cooler.winter.r.tau = priors %>% filter(.variable == "beta.cooler.winter.r") %>% pull(tau),
  prior.start.spring.r.mean = priors %>% filter(.variable == "beta.start.spring.r") %>% pull(mean),
  prior.start.spring.r.tau = priors %>% filter(.variable == "beta.start.spring.r") %>% pull(tau),
  prior.warm.dry.summer.r.mean = priors %>% filter(.variable == "beta.warm.dry.summer.r") %>% pull(mean),
  prior.warm.dry.summer.r.tau = priors %>% filter(.variable == "beta.warm.dry.summer.r") %>% pull(tau),
  prior.rain.on.snow.r.mean = priors %>% filter(.variable == "beta.rain.on.snow.r") %>% pull(mean),
  prior.rain.on.snow.r.tau = priors %>% filter(.variable == "beta.rain.on.snow.r") %>% pull(tau)
)


#  Initial values for N to avoid parent node erros
#  Indexing was off here
Nst <- array(NA_integer_, c(nherd, nyr, nage))

for (h in 1:nherd) {
  for (a in 1:nage) {
    Nst[h, first[h]:last[h], a] <- as.integer(
      max(
        c(n1[h, a], cdat$mu[cdat$herd == h], edat$mu[edat$herd == h])
      )
    )
  }
}

ipm_inits <- function() {
  list(
    N = Nst
  )
}


#  Model parameters to monitor
model_parms <- c(
  "N", "totNMF", "totN", "lambda",
  "S", "R", "R_adj",
  # "R.ad",

  "muS", "muR",
  "offset",
  "beta.dist.s",
  "beta.snow.depth.s",
  "beta.cooler.winter.s",
  "beta.start.spring.s",
  "beta.warm.dry.summer.s",
  "beta.rain.on.snow.s",
  "beta.dist.r",
  "beta.snow.depth.r",
  "beta.cooler.winter.r",
  "beta.start.spring.r",
  "beta.warm.dry.summer.r",
  "beta.rain.on.snow.r",
  "beta.type.otc.r", "beta.type.osc.r", "beta.type.mnka.r",

  # "sight_sd_yr_sg","sight_tau_yr_sg","sight_yr_sg",

  "p_mu", "p", "sight_est",
  "sexratio"
)
```

## Run IPM
```{r Run IPM, results='hide', message=FALSE, warning=FALSE}
nth <- 100
nbu <- 0
nch <- 3
nad <- 40000
nit <- 250000


# nth <- 1
# nbu <- 0
# nch <- 3
# nad <- 1000
# nit <- 5000

out <- jagsUI::jags(ipm_dat,
  inits = ipm_inits,
  model_parms,
  model.file = here::here("jags/Dist_Clim_IPM_rawVR.txt"),
  n.chains = nch,
  n.cores = nch,
  n.iter = nit,
  n.burnin = nbu,
  n.thin = nth,
  n.adapt = nad
)

saveRDS(out, file = here::here("jags/output/BCAB_CaribouIPM_posteriors.rds"))
```

## Plot draws
```{r plot draws, results='hide', message=FALSE, warning=FALSE}
draws <- out %>%
  gather_draws(
    muS, beta.dist.s, beta.snow.depth.s, beta.cooler.winter.s, beta.start.spring.s, beta.warm.dry.summer.s, beta.rain.on.snow.s,
    muR, beta.dist.r, beta.snow.depth.r, beta.cooler.winter.r, beta.start.spring.r, beta.warm.dry.summer.r, beta.rain.on.snow.r
  ) %>%
  mutate(type = "Posteriors") %>%
  rbind(out.priors %>%
    gather_draws(
      muS, beta.dist.s, beta.snow.depth.s, beta.cooler.winter.s, beta.start.spring.s, beta.warm.dry.summer.s, beta.rain.on.snow.s,
      muR, beta.dist.r, beta.snow.depth.r, beta.cooler.winter.r, beta.start.spring.r, beta.warm.dry.summer.r, beta.rain.on.snow.r
    ) %>%
    mutate(type = "Priors"))

draws %>%
  filter(!.variable %in% c("alpha.r", "alpha.s", "beta.season.r")) %>%
  ggplot(aes(x = .value, fill = type)) +
  geom_density(alpha = 0.5) +
  facet_wrap(vars(.variable), scales = "free") +
  geom_vline(xintercept = 0, linetype = "dashed")

draws %>%
  filter(!.variable %in% c("alpha.r", "alpha.s", "beta.season.r", "muS", "muR")) %>%
  ggplot(aes(x = .value, fill = type)) +
  geom_density(alpha = 0.5) +
  facet_wrap(vars(.variable)) +
  geom_vline(xintercept = 0, linetype = "dashed")
```
