==> scripts_r/3.results_summary.Rmd <==
---
title: "Asses pop trends with Climate-Disturbance IPM"
author: "Clayton T. Lamb"
date: '`r format(Sys.time(), "%d %B, %Y")`'
output: github_document
---


```{r render, eval=FALSE,include=FALSE}
# rmarkdown::render(here::here("jags", 'run_IPM.Rmd'),
#                   output_file =  "README.md")

knitr::purl(
  input = here::here("scripts_r", "3.results_summary.Rmd"),
  output = here::here("scripts_r", "3.results_summary.r")
)
```


## Load Data
```{r Load packages and data, results='hide', message=FALSE, warning=FALSE}
library(renv)
library(MCMCvis)
library(tidybayes)
library(ggmcmc)
library(boot)
library(cowplot)
library(hrbrthemes)
library(lme4)
library(tidymodels)
library(broom.mixed)
library(RColorBrewer)
library(ggeffects)
library(ggridges)
library(ggrepel)
library(tidylog)
library(gt)
library(patchwork)
library(ggtext)
library(rlang)
library(tidyverse)

##load helpers
source(here::here("scripts_r/helpers/gg_theme_custom.R"))

# Load data ---------------------------------------------------------------

## IPM Output
out <- readRDS(file = here::here("jags/output/BCAB_CaribouIPM_posteriors.rds"))

## IPM input to compare results
hd <- read.csv(here::here("data/clean/blueprint.csv"))
hn <- hd %>%
  dplyr::select(herd, herd_num)
afs <- read.csv(here::here("data/clean/survival.csv")) %>%
  arrange(herd) %>%
  left_join(hn, by = "herd")
afr <- read.csv(here::here("data/clean/recruitment.csv")) %>%
  arrange(herd) %>%
  left_join(hn, by = "herd")
counts <- read.csv(here::here("data/clean/counts.csv")) %>%
  arrange(herd) %>%
  left_join(hn, by = "herd")
trt <- read.csv(here::here("data/clean/treatments.csv")) %>%
  arrange(herd) %>%
  left_join(hn, by = "herd")
covar.names <- readRDS(here::here("data/clean/forIPM/covariate_names.rds"))
covariates <- read_csv(here::here("data/clean/forIPM/covariates_scaled.csv")) %>%
  dplyr::select(herd, year, !!covar.names, !!paste0(covar.names, "_scale"))
# covariates_scaled <- read_csv(here::here("data/clean/forIPM/covariates_scaled.csv")) %>%
#   dplyr::select(herd, year, !!paste0(covar.names,"_scale"))

labels <- read.csv(here::here("data/clean/labels.csv"))

#  Years of study
yr_df <- read.csv(here::here("data/clean/forIPM/yr_df.csv"))

##ecotype
ecotype <- read_csv(here::here("data/clean/ecotype.csv"))%>%
  select(herd=Herd, ecotype=ECCC_Recov_Grp)
```


```{r Check posteriors, eval=FALSE, message=FALSE, warning=FALSE, include=FALSE, results='hide'}
## check posteriors for convergence
mcmcplots::mcmcplot(out$samples, random = 100, par = "lambda")
# mcmcplots::mcmcplot(out$samples, random=100, par = c("sight_sd_yr_sg","sight_tau_yr_sg","sight_yr_sg"))
mcmcplots::mcmcplot(out$samples, random = 100, par = c("p_mu", "p", "sight_est"))

mcmcplots::mcmcplot(out$samples, random = 50, par = "totNMF")

mcmcplots::mcmcplot(out$samples, random = 100, par = "S")
mcmcplots::mcmcplot(out$samples, random = 100, par = "R")
mcmcplots::mcmcplot(out$samples, random = 100, par = "sexratio")

mcmcplots::mcmcplot(out$samples, random = 100, par = c("beta.type.otc.r", "beta.type.osc.r", "beta.type.mnka.r"))


mcmcplots::mcmcplot(out$samples, par = c(
  "muS", "muR",
  "offset",
  "beta.dist.s",
  "beta.snow.depth.s",
  "beta.start.spring.s",
  "beta.warm.dry.summer.s",
  "beta.rain.on.snow.s",
  "beta.dist.r",
  "beta.snow.depth.r",
  "beta.start.spring.r",
  "beta.warm.dry.summer.r",
  "beta.rain.on.snow.r"
)) ## look good
```


## ----Data housekeeping
```{r Data housekeeping, results='hide', message=FALSE, warning=FALSE}
## set up colors for plotting
display.brewer.pal(8, "Accent")
cols <- RColorBrewer::brewer.pal(8, "Accent")

### Change Narraway BC to Bearhole Redwillow
trt <- trt %>% mutate(herd = case_when(herd %in% "Narraway BC" ~ "Bearhole Redwillow", TRUE ~ herd))
hd <- hd %>% mutate(herd = case_when(herd %in% "Narraway BC" ~ "Bearhole Redwillow", TRUE ~ herd))
counts <- counts %>% mutate(herd = case_when(herd %in% "Narraway BC" ~ "Bearhole Redwillow", TRUE ~ herd))
afr <- afr %>% mutate(herd = case_when(herd %in% "Narraway BC" ~ "Bearhole Redwillow", TRUE ~ herd))
afs <- afs %>% mutate(herd = case_when(herd %in% "Narraway BC" ~ "Bearhole Redwillow", TRUE ~ herd))
covariates <- covariates %>% mutate(herd = case_when(herd %in% "Narraway BC" ~ "Bearhole Redwillow", TRUE ~ herd))
ecotype <- ecotype %>% mutate(herd = case_when(herd %in% "Narraway BC" ~ "Bearhole Redwillow", TRUE ~ herd))

# Pull out some values, summarize to demog data frame ---------------------

#### TREATMENT COMBOS
treatment.combos <- trt %>%
  filter(applied == 1) %>%
  group_by(herd, year) %>%
  summarize(
    trt = paste(
      gsub(" ", "", treatment), # Remove spaces
      collapse = "-" # Add - between treatments
    ),
    .groups = "drop"
  ) %>%
  dplyr::select(herd, yrs = year, trt)


## pull posterior draws, add in herd, year, and treatment to each herd-year
ndraws <- 1000
demog <- out %>%
  spread_draws(
    c(totNMF, totN, totAdults, S, R_adj, lambda, sexratio)[i, j],
    ndraws = ndraws
  ) %>%
  median_qi(.width = 0.9) %>%
  left_join(yr_df %>% rename(j = yr_idx), by = "j") %>%
  left_join(hd %>% dplyr::select(herd, i = herd_num), by = "i") %>%
  left_join(treatment.combos, by = c("herd", "yrs")) %>%
  mutate(trt = replace_na(trt, "Reference")) %>%
  rename(R = R_adj, R.lower = R_adj.lower, R.upper = R_adj.upper) %>%
  ungroup()


## filter herds to a period where not fully extirpated
extirpated.yr <- demog %>%
  filter(round(totNMF, 0) <= 1) %>%
  group_by(herd) %>%
  filter(yrs == min(yrs)) %>%
  ungroup()


herds <- unique(demog$herd)

demog.trim <- tibble()
for (i in 1:length(herds)) {
  if (herds[i] %in% extirpated.yr$herd) {
    yr <- extirpated.yr %>%
      dplyr::filter(herd == !!herds[i])
    a <- demog %>%
      dplyr::filter(herd == !!herds[i] & yrs <= yr$yrs)
  }

  if (!herds[i] %in% extirpated.yr$herd) {
    a <- demog %>%
      dplyr::filter(herd == !!herds[i])
  }

  demog.trim <- bind_rows(a, demog.trim)
}

demog <- demog.trim
rm(demog.trim)
rm(a)
rm(yr)


## ----Firstyr
raw.demog <- rbind(
  afr %>% dplyr::select(herd, year, est) %>% mutate(type = "Recruit"),
  afs %>% dplyr::select(herd, year, est) %>% mutate(type = "Surv"),
  counts %>% dplyr::select(herd, year, est = Est_CL) %>% mutate(type = "Count")
)

### Identify first year of demographic data for each herd
first.yr <- raw.demog %>%
  dplyr::select(herd, first.year = year) %>%
  distinct() %>%
  group_by(herd) %>%
  filter(first.year == min(first.year))

ggplot(first.yr, aes(x = first.year)) +
  stat_bin(aes(y = cumsum(..count..)), geom = "step") +
  theme_ipsum() +
  labs(
    x = "Year", y = "Cumulative count",
    title = "First year of demographic data for the 41 SMC herds",
    subtitle = "Cumulative count of herds being monitored through time"
  ) +
  theme(
    plot.title = element_text(size = 18),
    axis.title.y = element_text(size = 15),
    axis.title.x = element_text(size = 15),
    axis.text.x = element_text(size = 12),
    axis.text.y = element_text(size = 12),
    strip.text.x = element_text(size = 14),
    strip.text.y = element_text(size = 14),
    legend.text = element_text(size = 12),
    legend.title = element_text(size = 14),
    legend.position = "bottom"
  )


## shift S to y+1 for correlations
s.shift <- demog %>%
  mutate(year = yrs + 1) %>%
  select(herd, year, S, S.lower, S.upper)


demog.mod <- demog %>%
  filter(totNMF > 20) %>% ## modelling data that doesn't include functionally extirpated herds, demography gets unstable
  rename(year = yrs) %>%
  select(-S, -S.lower, -S.upper) %>%
  left_join(s.shift, by = c("herd", "year")) %>%
  left_join(covariates, by = c("herd", "year")) %>% ## add covariates
  ## keep only IPM results for each herd after first yr
  left_join(first.yr, by = "herd") %>%
  filter(year > first.year,
         lambda.upper-lambda.lower>0,
         !is.na(S))



ggplot(demog, aes(y = log(lambda), x = totNMF)) +
  geom_point(alpha = 0.3) +
  geom_vline(xintercept = 20, color = "red") +
  geom_hline(yintercept = 0, color = "grey", linetype = "dashed") +
  theme_ipsum() +
  theme(
    axis.title.x = element_text(size = 15),
    axis.title.y = element_text(size = 15),
    strip.text.x = element_text(size = 15),
    strip.text.y = element_text(size = 15)
  ) +
  labs(
    title = "Population growth unstable when subpopulation too small",
    subtitle = "remove data below functionally extirpated abundance (red line)",
    x = "Subopulation abundance",
    y = "Instanteous rate of increase (r)"
  )


ggplot(demog, aes(y = sexratio, x = yrs)) +
  geom_point() +
  theme_ipsum() +
  theme(
    axis.title.x = element_text(size = 15),
    axis.title.y = element_text(size = 15),
    strip.text.x = element_text(size = 15),
    strip.text.y = element_text(size = 15)
  ) +
  labs(
    x = "Year",
    y = "Sex ratio"
  ) +
  facet_wrap(vars(herd))
```

## Plot herd abundance
```{r Plot herd abundance, fig.height=11, fig.width=15, message=FALSE}
## Prep data and layout for plot

## treatments for plot,and where they go
trt.plot <- trt %>%
  filter(applied == 1) %>%
  left_join(demog %>% group_by(herd) %>% summarize(max = max(totNMF.upper))) %>%
  mutate(y = case_when(
    treatment %in% "reduce wolves" & herd %in% "South Selkirks" ~ 140,
    treatment %in% "transplant" & herd %in% "South Selkirks" ~ 165,
    treatment %in% "transplant" & herd %in% "Purcells South" ~ 180,
    treatment %in% "reduce moose" & herd %in% "Frisby-Boulder" ~ 200,
    treatment %in% "reduce wolves" & herd %in% "Barkerville" ~ max - (max * 0.1),
    treatment %in% "sterilize wolves" & herd %in% "Barkerville" ~ max - (max * 0.25),
    treatment %in% "reduce wolves" & herd %in% "Kennedy Siding" ~ max - (max * 0.1),
    treatment %in% "feed" & herd %in% "Kennedy Siding" ~ max - (max * 0.25),
    treatment %in% "reduce wolves" ~ max - (max * 0.1),
    treatment %in% "transplant" ~ max - (max * 0.15),
    treatment %in% "reduce moose" ~ max - (max * 0.2),
    treatment %in% "sterilize wolves" ~ max - (max * 0.25),
    treatment %in% "pen" ~ max - (max * 0.3),
    treatment %in% "feed" ~ max - (max * 0.35)
  ))

### a couple functions for the red boxes around herd names of extirpated
element_textbox_highlight <- function(..., hi.labels = NULL, hi.fill = NULL,
                                      hi.col = NULL, hi.box.col = NULL, hi.family = NULL) {
  structure(
    c(
      element_textbox(...),
      list(hi.labels = hi.labels, hi.fill = hi.fill, hi.col = hi.col, hi.box.col = hi.box.col, hi.family = hi.family)
    ),
    class = c("element_textbox_highlight", "element_textbox", "element_text", "element")
  )
}

element_grob.element_textbox_highlight <- function(element, label = "", ...) {
  if (label %in% element$hi.labels) {
    element$fill <- element$hi.fill %||% element$fill
    element$colour <- element$hi.col %||% element$colour
    element$box.colour <- element$hi.box.col %||% element$box.colour
    element$family <- element$hi.family %||% element$family
  }
  NextMethod()
}

## Herd Abundance
ggplot() +
  geom_ribbon(
    data = demog %>%
      ungroup() %>%
      left_join(labels, by = "herd") %>%
      mutate(herd = paste0(number_label, ".", herd, " (", human, "%)") %>%
        fct_reorder(.f = ., .x = number_label, .fun = max)),
    aes(x = yrs, y = totNMF, ymin = totNMF.lower, ymax = totNMF.upper), alpha = 0.4, color = NA, fill = cols[3]
  ) +
  geom_line(
    data = demog %>%
      ungroup() %>%
      left_join(labels, by = "herd") %>%
      mutate(herd = paste0(number_label, ".", herd, " (", human, "%)") %>%
        fct_reorder(number_label)),
    aes(x = yrs, y = totNMF, ymin = totNMF.lower, ymax = totNMF.upper), size = 1, color = cols[3]
  ) +
  geom_rug(
    data = raw.demog %>%
      ungroup() %>%
      left_join(labels, by = "herd") %>%
      mutate(herd = paste0(number_label, ".", herd, " (", human, "%)") %>%
        fct_reorder(number_label)),
    aes(x = year), sides = "t", length = unit(0.05, "npc"), alpha = 0.5
  ) +
  theme_ipsum() +
  theme(legend.position = "none") +
  facet_wrap(vars(herd), scales = "free_y", ncol = 7) +
  labs(x = "", y = "", title = "Abundance") +
  expand_limits(y = 0) +
  scale_x_continuous(
    limits = c(1970, 2026),
    breaks = seq(1970, 2020, by = 20)
  ) +
  theme(
    axis.title.x = element_text(size = 15),
    axis.title.y = element_text(size = 15),
    strip.text = element_textbox_highlight(
      size = 14,
      fill = "white", box.color = "white",
      halign = .5, linetype = 1, r = unit(0, "pt"), width = unit(1, "npc"),
      padding = margin(b = 2, t = 2, r = 2, l = 2), margin = margin(b = 8, t = 2),
      hi.fill = "firebrick", hi.box.col = "firebrick", hi.col = "white"
    ),
    legend.text = element_text(size = 13),
    legend.title = element_text(size = 15)
  ) +
  geom_point(
    data = counts %>%
      ungroup() %>%
      left_join(treatment.combos, by = c("herd", "year" = "yrs")) %>%
      left_join(labels, by = "herd") %>%
      mutate(herd = paste0(number_label, ".", herd, " (", human, "%)") %>%
        fct_reorder(number_label)),
    aes(x = year, y = Est_CL, color = !is.na(trt), fill = !is.na(trt)), size = 0.5, alpha = 0.7
  ) +
  geom_linerange(
    data = counts %>%
      ungroup() %>%
      left_join(treatment.combos, by = c("herd", "year" = "yrs")) %>%
      mutate(Est_CL.max = case_when(Est_CL.max > 5000 ~ 5000, TRUE ~ Est_CL.max)) %>%
      left_join(labels, by = "herd") %>% mutate(herd = paste0(number_label, ".", herd, " (", human, "%)") %>%
        fct_reorder(number_label)),
    aes(x = year, ymin = Est_CL.min, ymax = Est_CL.max, color = !is.na(trt)), alpha = 0.5
  ) +
  geom_point(
    data = trt.plot %>%
      ungroup() %>%
      left_join(labels, by = "herd") %>%
      mutate(herd = paste0(number_label, ".", herd, " (", human, "%)") %>%
        fct_reorder(number_label)),
    aes(x = year, y = y, group = treatment), size = 0.5
  ) +
  scale_color_manual(values = cols[-4]) +
  geom_text(
    data = trt.plot %>%
      ungroup() %>%
      distinct(herd, treatment, y) %>%
      mutate(t = str_remove(treatment, "reduce ") %>% str_sub(1, 1)) %>%
      left_join(labels, by = "herd") %>% mutate(herd = paste0(number_label, ".", herd, " (", human, "%)") %>%
        fct_reorder(number_label)),
    aes(label = t, x = 2025, y = y),
    direction = "y"
  ) +
  coord_cartesian(clip = "off")


ggsave(here::here("plots", "abundance.png"), width = 15, height = 11, bg = "white")
```


## Plot draws
```{r plot draws, results='hide', message=FALSE, warning=FALSE}
plot.draws <- out %>%
  gather_draws(
    beta.dist.s, beta.snow.depth.s, beta.cooler.winter.s, beta.start.spring.s, beta.warm.dry.summer.s, beta.rain.on.snow.s,
    beta.dist.r, beta.snow.depth.r, beta.cooler.winter.r, beta.start.spring.r, beta.warm.dry.summer.r, beta.rain.on.snow.r
  ) %>%
  mutate(
    type = str_sub(.variable, -1, -1)%>%str_to_upper(),
    var = str_sub(.variable, 6, -3)
  ) %>%
  select(var, .value, type)%>%
  ungroup%>%
  mutate(var=case_when(var=="dist"~"disturbance",
                       TRUE~str_replace_all(var,"\\."," ")))


plot.draws%>%
  group_by(var)%>%
  mutate(median=median(.value))%>% ##for ordering
ggplot(aes(x = .value, color = type, fill=type, y = fct_reorder(var, median))) +
  geom_density_ridges(alpha=0.4,rel_min_height = 0.01) +
  geom_vline(xintercept = 0, linetype = "dashed")+
  theme.custom+
  theme(legend.position = c(0.2,0.97),legend.direction="horizontal")+
  labs(x = "Estimate (standardized)", y = "Covariate", color = "Vital rate", fill = "Vital rate")

ggsave(here::here("plots", "AnalysisC_vitalrate_betas.png"), width = 6, height = 6, bg = "white")

```

```{r add lambda, eval=FALSE, message=FALSE, warning=FALSE, include=FALSE, results='hide'}
## add lambda

##
demog.draws <- out %>%
  spread_draws(
    c(totNMF, lambda, S)[i, j],
    ndraws = ndraws
  ) %>%
  left_join(yr_df %>% rename(j = yr_idx), by = "j") %>%
  left_join(hd %>% dplyr::select(herd, i = herd_num), by = "i") %>%
  left_join(treatment.combos, by = c("herd", "yrs")) %>%
  mutate(trt = replace_na(trt, "Reference")) %>%
  filter(trt == "Reference") %>%
  filter(totNMF > 20) %>% ## modelling data that doesn't include functionally extirpated herds, demography gets unstable
  rename(year = yrs) %>%
  left_join(covariates, by = c("herd", "year")) %>% ## add covariates
  ## keep only IPM results for each herd after first yr
  left_join(first.yr, by = "herd") %>%
  filter(year >= first.year) %>%
  ungroup()



lambda.mod.mm <- demog.draws %>%
  group_by(.draw) %>%
  do(tidy(lmer(log(lambda) ~ dist.hum.es_scale + snow.depth_scale + start.spring_scale + warm.dry.summer_scale + rain.on.snow_scale + (1 | herd), data = .))) %>%
  filter(!(term %in% "(Intercept)" | effect != "fixed")) %>%
  group_by(term) %>%
  summarise(
    .value = median(estimate),
    .lower = (quantile(estimate, 0.025)),
    .upper = (quantile(estimate, 0.975))
  ) %>%
  mutate(
    var = case_when(
      term == "dist.hum.es_scale" ~ "dist",
      TRUE ~ str_remove(term, "_scale")
    ),
    type = "lambda"
  ) %>%
  select(var, .value, .lower, .upper, type)


#
# yr.bin <- 1
# mod.dat <- demog.mod %>%
#   filter(trt == "Reference") %>%
#   group_by(herd, year = ceiling(year / yr.bin) * yr.bin) %>%
#   summarize(
#     year = paste(first(year) - yr.bin, first(year), sep = "-"),
#     r = mean(log(lambda), na.rm = TRUE),
#     S = mean(log(S), na.rm = TRUE)%>%exp(),
#     R = mean(log(R), na.rm = TRUE)%>%exp(),
#     dist = mean(dist.hum.es_scale, na.rm = TRUE),
#     snow.depth = mean(snow.depth_scale, na.rm = TRUE),
#     start.spring = mean(start.spring_scale, na.rm = TRUE),
#     warm.dry.summer = mean(warm.dry.summer_scale, na.rm = TRUE),
#     rain.on.snow = mean(rain.on.snow_scale, na.rm = TRUE)
#   ) %>%
#   ungroup()
#
# glm(r ~ dist + snow.depth + start.spring + warm.dry.summer + rain.on.snow, data = mod.dat)%>%
#   tidy()

out %>%
  gather_draws(
    beta.dist.s, beta.snow.depth.s, beta.start.spring.s, beta.warm.dry.summer.s, beta.rain.on.snow.s,
    beta.dist.r, beta.snow.depth.r, beta.start.spring.r, beta.warm.dry.summer.r, beta.rain.on.snow.r
  ) %>%
  median_qi() %>%
  mutate(
    type = str_sub(.variable, -1, -1),
    var = str_sub(.variable, 6, -3)
  ) %>%
  select(var, .value, .lower, .upper, type) %>%
  rbind(lambda.mod.mm) %>%
  ggplot(aes(x = .value, xmin = .lower, xmax = .upper, color = type, y = var)) +
  geom_linerange(alpha = 0.8, position = position_dodge(width = 0.3)) +
  geom_point(position = position_dodge(width = 0.3)) +
  geom_vline(xintercept = 0, linetype = "dashed") +
  labs(x = "Estimate (standardized)", y = "Covariate", color = "Demographic rate")
```

## Run simulations to assess C+D, D, C WITH ERROR
```{r Run sims, results='hide', message=FALSE, warning=FALSE}
ndraws <- 50


## load individual covariates
dist.hum <- readRDS(here::here("data/clean/forIPM/dist.hum.es.rds"))
snow.depth <- readRDS(here::here("data/clean/forIPM/snow.depth.rds"))
cooler.winter <- readRDS(here::here("data/clean/forIPM/cooler.winter.rds"))
start.spring <- readRDS(here::here("data/clean/forIPM/start.spring.rds"))
warm.dry.summer <- readRDS(here::here("data/clean/forIPM/warm.dry.summer.rds"))
rain.on.snow <- readRDS(here::here("data/clean/forIPM/rain.on.snow.rds"))

## fill 1970's covariate values for rain on snow
early.fill <- rain.on.snow[, 11:16] %>%
  rowwise() %>%
  mutate(m = mean(`11`:`16`))
rain.on.snow[, 1:10] <- early.fill$m

nherd <- max(demog$i)
nyr <- max(demog$j)

ipm.LP.betas.draws <- out %>%
  gather_draws(
    muS, beta.dist.s, beta.snow.depth.s, beta.cooler.winter.s, beta.start.spring.s, beta.warm.dry.summer.s, beta.rain.on.snow.s,
    muR, beta.dist.r, beta.snow.depth.r, beta.cooler.winter.r, beta.start.spring.r, beta.warm.dry.summer.r, beta.rain.on.snow.r, beta.type.osc.r,
    ndraws = ndraws
  ) %>%
  rename(median = .value) %>%
  ungroup()
sim.results <- tibble()

draws <- unique(ipm.LP.betas.draws$.draw)

for (draw in 1:length(draws)) {
  ipm.LP.betas.i <- ipm.LP.betas.draws %>%
    filter(.draw == draws[draw])

  # format is	 [hd,yr,age,sim]
  pred_N <- array(NA, c(nherd, nyr, 2, 3))

  pred_totCalves <- array(NA, c(nherd, nyr, 1, 3))
  pred_totAdults <- array(NA, c(nherd, nyr, 1, 3))
  pred_totN <- array(NA, c(nherd, nyr, 1, 3))

  pred_totCalvesMF <- array(NA, c(nherd, nyr, 1, 3))
  pred_totAdultsMF <- array(NA, c(nherd, nyr, 1, 3))
  pred_totNMF <- array(NA, c(nherd, nyr, 1, 3))

  pred_s.daily <- array(NA, c(nherd, nyr, 1, 3))
  pred_S <- array(NA, c(nherd, nyr, 1, 3))
  pred_R <- array(NA, c(nherd, nyr, 1, 3))

  ## turn covariates into array, 3, one for each sim
  ## sim 1==climate + disturbance
  ## sim 2==climate only
  ## sim 3==disturbance only
  yrs.baseline <- 1:20

  dist.hum.array <- array(NA, dim = c(nherd, nyr, 3))
  dist.hum.array[, , 1] <- dist.hum %>%
    as.data.frame() %>%
    unlist()
  dist.hum.array[, , 2] <- rowMeans(dist.hum.array[, yrs.baseline, 1])
  dist.hum.array[, , 3] <- dist.hum.array[, , 1]

  # dist.hum.array[, 51, 1] == dist.hum.array[, 51, 3]
  # dist.hum.array[, 51, 1] != dist.hum.array[, 51, 2]

  snow.depth.array <- array(NA, dim = c(nherd, nyr, 3))
  snow.depth.array[, , 1] <- snow.depth %>%
    as.data.frame() %>%
    unlist()
  snow.depth.array[, , 2] <- snow.depth.array[, , 1]
  snow.depth.array[, , 3] <- rowMeans(snow.depth.array[, yrs.baseline, 1])
  
  cooler.winter.array <- array(NA, dim = c(nherd, nyr, 3))
  cooler.winter.array[, , 1] <- cooler.winter %>%
    as.data.frame() %>%
    unlist()
  cooler.winter.array[, , 2] <- cooler.winter.array[, , 1]
  cooler.winter.array[, , 3] <- rowMeans(cooler.winter.array[, yrs.baseline, 1])

  start.spring.array <- array(NA, dim = c(nherd, nyr, 3))
  start.spring.array[, , 1] <- start.spring %>%
    as.data.frame() %>%
    unlist()
  start.spring.array[, , 2] <- start.spring.array[, , 1]
  start.spring.array[, , 3] <- rowMeans(start.spring.array[, yrs.baseline, 1])

  warm.dry.summer.array <- array(NA, dim = c(nherd, nyr, 3))
  warm.dry.summer.array[, , 1] <- warm.dry.summer %>%
    as.data.frame() %>%
    unlist()
  warm.dry.summer.array[, , 2] <- warm.dry.summer.array[, , 1]
  warm.dry.summer.array[, , 3] <- rowMeans(warm.dry.summer.array[, yrs.baseline, 1])

  rain.on.snow.array <- array(NA, dim = c(nherd, nyr, 3))
  rain.on.snow.array[, , 1] <- rain.on.snow %>%
    as.data.frame() %>%
    unlist()
  rain.on.snow.array[, , 2] <- rain.on.snow.array[, , 1]
  rain.on.snow.array[, , 3] <- rowMeans(rain.on.snow.array[, yrs.baseline, 1])

  ##########################################################################
  # #Simulate climate+dist, climate-only, and disturbance-only scenarios
  # pred_N[, 1, 1, ] <- out$mean$N[, 1, 1]
  # pred_N[, 1, 2, ] <- out$mean$N[, 1, 2]
  #
  # # format is	 [hd,yr,sim]
  # pred_s.daily[, 1, 1, ] <- out$mean$S[, 2]^(1 / 365) ## first herd-specific S available, transform to daily
  # pred_S[, 1, 1, ] <- pred_s.daily[, 1, 1, ]^365
  # pred_R[, 1, 1, ] <- out$mean$R[, 2]
  #
  # pred_totCalves[, 1, 1, ] <- out$mean$N[, 1, 1]
  # pred_totAdults[, 1, 1, ] <- out$mean$N[, 1, 2]
  # pred_totN[, 1, 1, ] <- out$mean$N[, 1, 1] + out$mean$N[, 1, 2]
  # #
  ## when no first year for all
  pred_N[, 1, 1, ] <- 25
  pred_N[, 1, 2, ] <- 150

  # format is	 [hd,yr,sim]
  pred_s.daily[, 1, 1, ] <- 0.85^(1 / 365) ## first herd-specific S available, transform to daily
  pred_S[, 1, 1, ] <- 0.85
  pred_R[, 1, 1, ] <- 0.18

  pred_totCalves[, 1, 1, ] <- pred_N[, 1, 1, ]
  pred_totAdults[, 1, 1, ] <- pred_N[, 1, 2, ]
  pred_totN[, 1, 1, ] <- pred_totCalves[, 1, 1, ] + pred_totAdults[, 1, 1, ]

  #  Apply sex ratio to get full population size
  pred_totAdultsMF[, 1, 1, ] <- pred_totAdults[, 1, 1, ] / 0.66
  pred_totCalvesMF[, 1, 1, ] <- pred_totCalves[, 1, 1, ] / 0.5
  pred_totNMF[, 1, 1, ] <- pred_totAdultsMF[, 1, 1, ] + pred_totCalvesMF[, 1, 1, ]

  for (h in 1:nherd) {
    for (yr in 2:nyr) {
      for (sim in 1:3) {
        pred_s.daily[h, yr, 1, sim] <- as.numeric(ipm.LP.betas.i %>% filter(.variable == "muS") %>% pull(median) + # intercept (baseline survival)
          ipm.LP.betas.i %>% filter(.variable == "beta.dist.s") %>% pull(median) * dist.hum.array[h, yr - 1, sim] + ## note dist.hum here and dist in IPM. Same data, just different notation
          ipm.LP.betas.i %>% filter(.variable == "beta.snow.depth.s") %>% pull(median) * snow.depth.array[h, yr - 1, sim] +
          ipm.LP.betas.i %>% filter(.variable == "beta.cooler.winter.s") %>% pull(median) * cooler.winter.array[h, yr - 1, sim] +
          ipm.LP.betas.i %>% filter(.variable == "beta.start.spring.s") %>% pull(median) * start.spring.array[h, yr - 1, sim] +
          ipm.LP.betas.i %>% filter(.variable == "beta.warm.dry.summer.s") %>% pull(median) * warm.dry.summer.array[h, yr - 1, sim] +
          ipm.LP.betas.i %>% filter(.variable == "beta.rain.on.snow.s") %>% pull(median) * rain.on.snow.array[h, yr - 1, sim]) %>% plogis()
        # s.daily to annual survival
        pred_S[h, yr, 1, sim] <- pred_s.daily[h, yr, 1, sim]^365

        pred_R[h, yr, 1, sim] <- as.numeric(ipm.LP.betas.i %>% filter(.variable == "muR") %>% pull(median) + # intercept (baseline recruitment)
          ipm.LP.betas.i %>% filter(.variable == "beta.dist.r") %>% pull(median) * dist.hum.array[h, yr, sim] + ## note dist.hum here and dist in IPM. Same data, just different notation
          ipm.LP.betas.i %>% filter(.variable == "beta.snow.depth.r") %>% pull(median) * snow.depth.array[h, yr, sim] +
          ipm.LP.betas.i %>% filter(.variable == "beta.cooler.winter.r") %>% pull(median) * cooler.winter.array[h, yr, sim] +
          ipm.LP.betas.i %>% filter(.variable == "beta.start.spring.r") %>% pull(median) * start.spring.array[h, yr, sim] +
          ipm.LP.betas.i %>% filter(.variable == "beta.warm.dry.summer.r") %>% pull(median) * warm.dry.summer.array[h, yr, sim] +
          ipm.LP.betas.i %>% filter(.variable == "beta.rain.on.snow.r") %>% pull(median) * rain.on.snow.array[h, yr, sim] +
          ipm.LP.betas.i %>% filter(.variable == "beta.type.osc.r") %>% pull(median)) %>% plogis()

        #  SubAdults and Adults: Binomial, dbin(p,n), 0 < p < 1 and n ≥ 1
        pred_N[h, yr, 2, sim] <- pred_S[h, yr - 1, 1, sim] * (pred_N[h, yr - 1, 2, sim] + pred_N[h, yr - 1, 1, sim])
        pred_totAdults[h, yr, 1, sim] <- pred_N[h, yr, 2, sim]
        pred_totAdultsMF[h, yr, 1, sim] <- pred_totAdults[h, yr, 1, sim] / 0.66

        # Calves
        pred_N[h, yr, 1, sim] <- (pred_totAdultsMF[h, yr, 1, sim] * pred_R[h, yr, 1, sim]) * 0.5


        #  Totals in each year - females only
        pred_totCalves[h, yr, 1, sim] <- pred_N[h, yr, 1, sim]
        pred_totN[h, yr, 1, sim] <- pred_totCalves[h, yr, 1, sim] + pred_totAdults[h, yr, 1, sim]

        #  Apply sex ratio to get full population size
        pred_totCalvesMF[h, yr, 1, sim] <- pred_totCalves[h, yr, 1, sim] / 0.5
        pred_totNMF[h, yr, 1, sim] <- pred_totAdultsMF[h, yr, 1, sim] + pred_totCalvesMF[h, yr, 1, sim]
      }
    }
  }
  sim.results <- sim.results %>%
    bind_rows(pred_totNMF[, , , 1] %>%
      as_tibble() %>%
      mutate(
        sim = "Climate+Disturbance",
        herd_num = row_number()
      ) %>%
      bind_rows(pred_totNMF[, , , 2] %>%
        as_tibble() %>%
        mutate(
          sim = "Climate",
          herd_num = row_number()
        )) %>%
      bind_rows(pred_totNMF[, , , 3] %>%
        as_tibble() %>%
        mutate(
          sim = "Disturbance",
          herd_num = row_number()
        )) %>%
      mutate(draw = draws[draw]))
}

write_csv(sim.results, here::here("tables", "sim.results.csv"))

sim.results <- read_csv(here::here("tables", "sim.results.csv"))

sim.results %>%
  pivot_longer(-c("herd_num", "sim", "draw")) %>%
  mutate(yr_idx = parse_number(name)) %>%
  left_join(yr_df, by = "yr_idx") %>%
  left_join(hd, by = "herd_num") %>%
  select(draw, sim, herd, yrs, value) %>%
  group_by(herd, sim, yrs) %>%
  summarise(value = median(value)) %>%
  ggplot(aes(x = yrs, y = value, color = sim, group = sim)) +
  geom_line() +
  facet_wrap(vars(herd), scales = "free")+
  scale_y_continuous(limits = c(0, NA))

## sim 1==climate + disturbance
## sim 2==climate only
## sim 3==disturbance only
sim.results %>%
  pivot_longer(-c("herd_num", "sim", "draw")) %>%
  mutate(yr_idx = parse_number(name)) %>%
  left_join(yr_df, by = "yr_idx") %>%
  left_join(hd, by = "herd_num") %>%
  select(draw, sim, herd, yrs, value) %>%
  group_by(draw, sim, yrs) %>%
  summarise(value = sum(value)) %>%
  ungroup() %>%
  arrange(yrs, sim, draw) %>%
  ggplot() +
  geom_line(aes(x = yrs, y = value, color = sim, fill = sim, group = interaction(draw, sim)), alpha = 0.1) +
  scale_y_continuous(limits = c(0, NA)) +
  theme_ipsum() +
  labs(x = "Year", y = "SMC abundance")


sim.results %>%
  pivot_longer(-c("herd_num", "sim", "draw")) %>%
  mutate(yr_idx = parse_number(name)) %>%
  left_join(yr_df, by = "yr_idx") %>%
  left_join(hd, by = "herd_num") %>%
  left_join(ecotype%>%distinct(), by="herd")%>%
  group_by(ecotype, draw, sim, yrs) %>%
  summarise(value = sum(value)) %>%
  ungroup() %>%
  group_by(ecotype,sim, yrs) %>%
  summarise(
    median = median(value),
    lower = quantile(value, 0.05),
    upper = quantile(value, 0.95)
  ) %>%
  ungroup%>%
  mutate(ecotype=fct_relevel(ecotype,"Southern Group","Central Group", "Northern Group"))%>%
  ggplot(aes(x = yrs, y = median, ymin = lower, ymax = upper, color = sim, fill = sim)) +
  geom_ribbon(alpha = 0.2, color = NA) +
  geom_line() +
  scale_y_continuous(limits = c(0, NA)) +
  theme.custom+
  labs(x = "Year", y = "SMC abundance", color="",fill="")+
  facet_wrap(vars(ecotype), scales="free")
ggsave(here::here("plots", "AnalysisC_climate.sim.byECCC.png"), width = 10, height = 6, bg = "white")

sim.results %>%
  pivot_longer(-c("herd_num", "sim", "draw")) %>%
  mutate(yr_idx = parse_number(name)) %>%
  left_join(yr_df, by = "yr_idx") %>%
  left_join(hd, by = "herd_num") %>%
  select(draw, sim, herd, yrs, value) %>%
  group_by(draw, sim, yrs) %>%
  summarise(value = sum(value)) %>%
  ungroup() %>%
  group_by(sim, yrs) %>%
  summarise(
    median = median(value),
    lower = quantile(value, 0.05),
    upper = quantile(value, 0.95)
  ) %>%
  ggplot(aes(x = yrs, y = median, ymin = lower, ymax = upper, color = sim, fill = sim)) +
  geom_ribbon(alpha = 0.2, color = NA) +
  geom_line() +
  scale_y_continuous(limits = c(0, NA)) +
  theme.custom+
  labs(x = "Year", y = "SMC abundance", color="",fill="")

ggsave(here::here("plots", "AnalysisC_climate.sim.png"), width = 5, height = 6, bg = "white")


sim.results %>%
  pivot_longer(-c("herd_num", "sim", "draw")) %>%
  mutate(yr_idx = parse_number(name)) %>%
  left_join(yr_df, by = "yr_idx") %>%
  left_join(hd, by = "herd_num") %>%
  select(draw, sim, herd, yrs, value) %>%
  group_by(draw, sim, yrs) %>%
  summarise(value = sum(value)) %>%
  group_by(draw, sim) %>%
  arrange(yrs)%>%
  mutate(lambda=value/lag(value))%>%
  slice(-1)%>%
  summarise(lambda=exp(mean(log(lambda))))%>%
  group_by(sim)%>%
 summarise(
    median = median(lambda),
    lower = quantile(lambda, 0.05),
    upper = quantile(lambda, 0.95)
  )


# pred_totNMF[13, , , 1]
# pred_totNMF[13, , , 2]
# pred_totNMF[13, , , 3]
#
# pred_S[13, , , 1] %>% mean()
# pred_S[13, , , 2] %>% mean()
# pred_S[13, , , 3] %>% mean()
#
# pred_R[13, , , 1] %>% mean()
# pred_R[13, , , 2] %>% mean()
# pred_R[13, , , 3] %>% mean()
```

```{r explore lambda relationships, fig.height=11, fig.width=15, message=FALSE}
cor(demog.mod$lambda, demog.mod$R)
cor(demog.mod$lambda, demog.mod$S)
cor(demog.mod$R, demog.mod$S)

plot(demog.mod$S, demog.mod$lambda)
plot(demog.mod$R, demog.mod$lambda)
plot(demog.mod$R, demog.mod$S)

plot(demog$S, demog$lambda)


lmer(lambda ~ S + (1 | herd), data = demog.mod) %>% summary()
lmer(lambda ~ R + (1 | herd), data = demog.mod) %>% summary()


ggplot(demog.mod, aes(x = S, y = lambda, group = herd, color = herd)) +
  geom_point() +
  geom_smooth(method = "lm", se = FALSE) +
  theme(legend.position = "none")

ggplot(demog.mod, aes(x = R, y = lambda, group = herd, color = herd)) +
  geom_point() +
  geom_smooth(method = "lm", se = FALSE) +
  theme(legend.position = "none")

plot.dat <- demog.mod %>%
  group_by(herd) %>%
  arrange(year) %>%
  ## calculate alternate population growth rates (over 3 and 5 year)
  mutate(
    herd = as.factor(herd),
    trt = as.factor(trt),
    weight.i = 1 / ((lambda.upper - lambda.lower) / lambda),
    weights = weight.i / mean(weight.i),
    lambda_3 = (totNMF / lag(totNMF, 3))^(1 / 3),
    lambda_5 = (totNMF / lag(totNMF, 5))^(1 / 5)
  ) %>%
  ungroup()


plot.dat %>%
  filter(trt == "Reference") %>%
  ggplot(aes(x = dist.hum.es, y = lambda)) +
  geom_point() +
  geom_hline(yintercept = 1, linetype = "dashed") +
  geom_smooth()

plot.dat %>%
  filter(trt == "Reference") %>%
  ggplot(aes(x = dist.hum.es, y = lambda_5)) +
  geom_point() +
  geom_hline(yintercept = 1, linetype = "dashed") +
  geom_smooth()



library(mgcv)
m1 <- mgcv::gam(
  lambda_3 ~
    s(snow.depth) +
    s(start.spring) +
    s(warm.dry.summer) +
    s(rain.on.snow) +
    s(dist.hum.es, k = 5) +

    s(herd, bs = "re") +

    s(snow.depth, herd, bs = "re") +
    s(warm.dry.summer, herd, bs = "re") +
    s(dist.hum.es, herd, bs = "re") +
    s(start.spring, herd, bs = "re") +
    s(rain.on.snow, herd, bs = "re"),
  data = plot.dat
)


summary(m1)$r.sq

plot(m1, page = 1, scheme = 2)

m1.n <- mgcv::gam(
  totNMF ~
    s(snow.depth) +
    s(start.spring) +
    s(warm.dry.summer) +
    s(rain.on.snow) +
    s(dist.hum.es, k = 5) +

    s(herd, bs = "re") +

    s(snow.depth, herd, bs = "re") +
    s(warm.dry.summer, herd, bs = "re") +
    s(dist.hum.es, herd, bs = "re") +
    s(start.spring, herd, bs = "re") +
    s(rain.on.snow, herd, bs = "re"),
  # data = plot.dat
  data = plot.dat %>% filter(!herd %in% c("Takla", "Itcha-Ilgachuz"))
)


summary(m1.n)$r.sq

plot(m1.n, page = 1, scheme = 2)




### plot dist vs lambda summarized

yr.bin <- 10
mod.dat <- demog.mod %>%
  filter(trt == "Reference") %>%
  group_by(herd, year = ceiling(year / yr.bin) * yr.bin) %>%
  summarize(
    year = paste(first(year) - yr.bin, first(year), sep = "-"),
    r = mean(log(lambda), na.rm = TRUE),
    S = mean(S, na.rm = TRUE),
    R = mean(R, na.rm = TRUE),
    N = median(totNMF, na.rm = TRUE),
    dist = mean(dist.hum.es, na.rm = TRUE),
    dist.all = mean(dist.all, na.rm = TRUE),
    snow.depth = mean(snow.depth, na.rm = TRUE),
    start.spring = mean(start.spring, na.rm = TRUE),
    warm.dry.summer = mean(warm.dry.summer, na.rm = TRUE),
    rain.on.snow = mean(rain.on.snow, na.rm = TRUE),
    n=n(),
  ) %>%
  ungroup() %>%
  filter(n>=5,
    #!herd%in%c("Maligne", "Banff", "Tonquin", "Brazeau"),
    r < 0.2,
    r > (-0.2)
  )

lambda.dist <- ggplot(data=mod.dat,aes(x = dist, y = r)) +
  geom_point(alpha=0.5, aes(size=N)) +
  geom_smooth(method = "loess",aes(color="All"),se=FALSE, alpha=0.5) +
  geom_smooth(data=mod.dat%>%filter(!herd%in%c("Maligne", "Banff", "Tonquin", "Brazeau")),method = "loess",aes(color="w/o AB Parks"),se=FALSE, alpha=0.5, linetype="dashed") +
  geom_hline(yintercept = 0, linetype = "dashed") +
  labs(
    x = "Disturbance (proportion of herd area)",
    y = "Instantaneous growth rate",
    color="Data"
  )+
  scale_color_manual(values=c("All" = "red", "w/o AB Parks" = "forestgreen"))+
  scale_size_continuous(breaks = c(100, 500, 1000))+
  theme.custom

cor(mod.dat$r, mod.dat$R)
cor(mod.dat$r, mod.dat$S)

plot(mod.dat$R, mod.dat$r)
plot(mod.dat$S, mod.dat$r)

library(lme4)
lmer(r ~ S + (1 + S | herd), data = mod.dat) %>% summary()
lmer(r ~ R + (1 + R | herd), data = mod.dat) %>% summary()



yr.bin <- 10
mod.dat.ch <- demog.mod %>%
  filter(trt == "Reference") %>%
  select(herd, year, lambda, totNMF) %>%
  left_join(read_csv(here::here("data/spatial/ch.dist.csv")) %>% mutate(herd = str_trim(herd)), by = c("herd", "year")) %>%
  group_by(herd, CH_type, year = ceiling(year / yr.bin) * yr.bin) %>%
  summarize(
    year = paste(first(year) - yr.bin, first(year), sep = "-"),
    r = exp(mean(log(lambda), na.rm = TRUE)),
    N = min(totNMF, na.rm = TRUE),
    dist = mean(dist.hum, na.rm = TRUE),
    n=n()
  )%>%
  filter(n>=5)






ggplot(
  mod.dat.ch %>% pivot_wider(names_from = "CH_type", values_from = dist),
  aes(x = HEWSR, y = `LEWSR/Matrix`)
) +
  geom_point()


mod.dat.ch.50 <- mod.dat.ch %>%
  pivot_wider(names_from = "CH_type", values_from = dist) %>%
  filter(N > 50, !herd %in% c("Maligne", "Tonquin", "Narraway AB")) ## no CH delineated

ggplot(
  mod.dat.ch.50,
  aes(x = HEWSR, y = `LEWSR/Matrix`, size = r >= 1, color = r >= 1)
) +
  geom_point(alpha = 0.4)

ggplot(
  mod.dat.ch.50,
  aes(x = HEWSR, y = `LEWSR/Matrix`, size = r, color = r)
) +
  geom_point(alpha = 0.5)

ggplot(
  mod.dat.ch.50,
  aes(x = HEWSR, y = `LEWSR/Matrix`, size = r >= 0.99, color = r >= 0.99)
) +
  geom_rect(aes(xmin = 0 - 0.005, xmax = 0.05, ymin = 0 - 0.01, ymax = 0.25),
    fill = "transparent", color = "grey", size = 1, linetype = "dashed"
  ) +
  geom_point(alpha = 0.5) +
  labs(
    x = "High elevation disturbance (proportion)",
    y = "Matrix disturbance (proportion)",
    size = "Stable?",
    color = "Stable?"
  )



ggplot(
  mod.dat.ch.50,
  aes(x = HEWSR, y = `LEWSR/Matrix`, size = N, color = N)
) +
  geom_point(alpha = 0.4)


mod.dat.ch.50 %>%
  ungroup() %>%
  mutate(
    HEWSR.cat = cut_width(HEWSR, 0.05),
    Matrix.cat = cut_width(`LEWSR/Matrix`, 0.1)
  ) %>%
  group_by(HEWSR.cat, Matrix.cat) %>%
  summarise(
    lambda = exp(log(r) %>% mean()),
    n = n(),
    ss = sum(r >= 1) / n
  ) %>%
  ggplot() +
  geom_tile(aes(x = HEWSR.cat, y = Matrix.cat, fill = ss)) +
  geom_text(aes(x = HEWSR.cat, y = Matrix.cat, label = n), color = "white") +
  theme_ipsum() +
  labs(
    x = "High elevation disturbance (proportion)",
    y = "Matrix disturbance (proportion)"
  )

ss.grid <- mod.dat.ch.50 %>%
  ungroup() %>%
  mutate(
    HEWSR.cat = cut_width(HEWSR, 0.1, boundary = 0)%>%str_remove_all("\\(|\\[|\\]")%>%str_replace(",","-"),
    Matrix.cat = cut_width(`LEWSR/Matrix`, 0.15, boundary = 0)%>%str_remove_all("\\(|\\[|\\]")%>%str_replace(",","-"),
  ) %>%
  group_by(HEWSR.cat, Matrix.cat) %>%
  summarise(
    lambda = exp(log(r) %>% mean()),
    n = n(),
    ss.count = sum(r >= 0.99),
    ss = ss.count / n
  ) %>%
  ggplot() +
  geom_tile(aes(x = HEWSR.cat, y = Matrix.cat, fill = ss)) +
  geom_text(aes(x = HEWSR.cat, y = Matrix.cat, label = paste0(ss.count, "/", n)), color = "white") +
  theme.custom+
  theme(legend.position = "bottom")+
  labs(
    x = "High elevation disturbance (proportion)",
    y = "Matrix disturbance (proportion)",
    fill="Self-sustaining \n(proportion)"
  )

plot1 <-mod.dat.ch.50 %>%
  ungroup() %>%
  filter(r >= 0.99)%>%
  ggplot(aes(x = HEWSR)) + 
  geom_histogram(color = "black", fill = "white",binwidth=0.05) + 
  theme_void()+
  theme(axis.text.x=element_text(size=10))

plot1 <-mod.dat.ch.50 %>%
  ungroup() %>%
  ggplot(aes(x = HEWSR, fill=r>=0.99)) + 
  geom_density(binwidth=0.05, color="black", alpha=0.4) + 
  scale_fill_manual(values=c("black","dodgerblue"))+
  theme_void()+
  theme(axis.text.x=element_text(size=10),
        legend.position = "none")

plot2 <-mod.dat.ch.50 %>%
  ungroup() %>%
  filter(r >= 0.99)%>%
  ggplot(aes(x = `LEWSR/Matrix`)) + 
  geom_histogram(color = "black", fill = "white",binwidth=0.05)+
  coord_flip()+
  theme_void()+
  theme(axis.text.y=element_text(size=10))

plot2 <-mod.dat.ch.50 %>%
  ungroup() %>%
  ggplot(aes(x = `LEWSR/Matrix`, fill=r>=0.99)) + 
  geom_density(binwidth=0.05, color="black", alpha=0.4) + 
  scale_fill_manual(values=c("black","dodgerblue"))+
  theme_void()+
    coord_flip()+
  theme(axis.text.y=element_text(size=10),
        legend.position = "none")

ss.grid  <- plot1 + plot_spacer() + ss.grid + plot2 + 
  plot_layout(
    ncol = 2, 
    nrow = 2, 
    widths = c(4, 1),
    heights = c(1, 4)
  ) 

lambda.dist + ss.grid
ggsave(here::here("plots", "self.sustaining.prop.png"), width = 10, height = 6, bg = "white")



mod.dat.ch.50 %>%
  ungroup() %>%
  mutate(
    HEWSR.cat = cut_width(HEWSR, 0.05),
    Matrix.cat = cut_width(`LEWSR/Matrix`, 0.1)
  ) %>%
  group_by(HEWSR.cat, Matrix.cat) %>%
  summarise(
    lambda = exp(log(r) %>% mean()),
    n = n(),
    ss = sum(r >= 1) / n
  ) %>%
  ggplot() +
  geom_tile(aes(x = HEWSR.cat, y = Matrix.cat, fill = lambda)) +
  theme_ipsum() +
  labs(
    x = "High elevation disturbance (proportion)",
    y = "Matrix disturbance (proportion)"
  )
```
